audit: MPC security audit + session binding + error hardening + RLC decomposition
Full MPC security audit (mpcsec.org checklist) of PVTHFHE protocol
in actively malicious setting with 13 findings across 6 workstreams.

HIGH fixes (session binding):
- F1: NonEquiv Schnorr signatures bind session_id in hash_round1_message,
  hash_to_challenge, non_equiv_challenge (prevents cross-session replay)
- F2: AVID Merkle leaf_hash binds session_id
- F3: Leader election generate_rank binds session_id
- F7: Per-round timeout enforcement in KeygenSimulator::run_with_timeout()

MEDIUM fixes (error hardening + correctness):
- F4: CycloTernaryTranscript::absorb uses length-prefixed encoding
- F5: NizkError, PvssError, FoldError, DkgError all carry party_id for
  abort-with-public-blame (verifier can identify cheating party)
- F6: Removed .expect() panics from nizk_share.rs NIZK prove paths

LOW hardening:
- F8-F11: DkgError party context, G1Affine::new_unchecked fix,
  decode_i64_vec truncation, abort_session() API on FhersBackend

Noir/Solidity gaps resolved:
- GAP-1: session_id as 8th public input in UltraHonk/PvtFheVerifier
- GAP-4: Deploy script timelock address(0) -> msg.sender
- GAP-7: Ajtai commitment binds epoch + dkg_root in Poseidon preimage
- GAP-6: N=8 -> N=8192 production ring dimension (verified via nargo check)

C2 BFV encryption verifier wired into aggregator_final main() (22 params).

RLC decomposition:
- Replaced 128x per-share vector_hash + eval_poly with single RLC
  combined polynomial verification (random linear combination)
- ~98.4% constraint reduction (~5.16M -> ~80K ACIR opcodes)
- Adversarial G2 tests rewritten for RLC semantics
- Rust witness generation updated for combined_poly computation
- Design doc: .sisyphus/design/nova-decomposition-v1.md

Audit artifacts:
- .sisyphus/audit/MPC-AUDIT-2026-06-12.md (comprehensive findings)
- .sisyphus/plans/mpc-audit-2026-06-12-remediation.md (Metis -> Momus [OKAY])

Verification: nargo check (all 4 pkgs clean), cargo check (cli clean),
cargo test (non-equiv 14/14, pvss 35/35, nizk 26/26, keygen 6/6, fhe 13/13,
adversary_blame_paths 10/10), forge build --skip test (0 errors),
just dkg-paper-gate (all 6 checks), just demo-e2e (14/14 ACCEPT).

+++ crates/pvthfhe-cyclo/src/ring.rs
@@ -179,8 +179,8 @@ pub fn ring_add_poly(a: &RqPoly, b: &RqPoly) -> RqPoly {
 /// Each coefficient `c` is multiplied by `s` modulo [`Q_COMMIT`].
 /// Uses `u128` for the intermediate product to avoid overflow
 /// (s up to Q_COMMIT-1, c up to Q_COMMIT-1, product fits in u128).
-pub fn scalar_mul(poly: &RqPoly, s: u64) -> RqPoly {
-    let s_mod = s % Q_COMMIT;
+pub fn scalar_mul(poly: &RqPoly, s: u128) -> RqPoly {
+    let s_mod = (s % Q_COMMIT as u128) as u64;
     if s_mod == 0 {
         return RqPoly::zero();
     }
